# src/main/resources/specs/collector-manager-api-spec.yaml
openapi: 3.0.3
info:
  title: Recolnat
  description: |-
    Le réseau national des collections naturalistes 
    est une infrastructure de recherche française dont l'objectif est 
    de produire un corpus de données basé sur les collections d'histoire naturelle de France, 
    permettant de valoriser les recherches au service de l'étude de la géo- et de la biodiversité actuelle et passée.
    Liens utiles:
    - [RECOLNAT](https://fr.wikipedia.org/wiki/RECOLNAT)
  version: "1.0"
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@mnhn.fr
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: https://petstore3.swagger.io/api/v3
...
security:
  - collection-oidc: [ ] # <--- Use the same name as specified in securitySchemes*, here without scope
paths:
...
  /v1/specimens:
    get:
      operationId: retrieveAllSpecimen
      tags:
        - Specimen
      summary: Retrieve All Specimen
      parameters:
        - name: q
          required: false
          in: query
          description: String to look for in scientific name or catalog number
          schema:
            type: string
        - name: state
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/OperationType'
        - name: currentDetermination
          in: query
          description: >
            display specimen with currentDetermination:
             * `true` - filter on identification current
             * `false` - all identification
          schema:
            default: true
            type: boolean
        - name: levelType
          in: query
          description: >
            display specimen with taxon levelType defines:
             * `true` -  filter on taxon levelType equals MASTER
             * `false` - none filters on levelType taxon
          schema:
            default: true
            type: boolean
        - name: page
          required: true
          in: query
          schema:
            type: integer
            format: int32
            minimum: 0
          description: number of page to retrieve
          example:
            id: 1
        - name: size
          required: true
          in: query
          schema:
            type: integer
            format: int32
          description: number of elements per page to retrieve
          example:
            id: 10
        - name: columnSort
          required: false
          in: query
          description: >
            Sort type:
              * `modifiedAt` - Default
              * `catalogNumber` 
              * `collection` 
              * `name` 
              * `state` 
          schema:
            type: string
            default: modifiedAt
            enum:
              - state
              - catalogNumber
              - collection
              - name
              - modifiedAt
        - name: typeSort
          in: query
          description: >
            Sort order:
             * `ASC` - Ascending, from A to Z
             * `DESC` - Descending, from Z to A
          schema:
            default: DESC
            type: string
            enum:
              - ASC
              - DESC
        - name: institution_id
          required: false
          in: query
          description: Id of the institution of the specimens
          schema:
            type: string
            format: uuid
          example: 386ff0b0-a6cc-11ec-b909-0242ac120002
        - name: collection_id
          required: false
          in: query
          description: Id of the collection of the specimens
          schema:
            type: string
            format: uuid
          example: 386ff0b0-a6cc-11ec-b909-0242ac120002
        - name: collection_code
          required: false
          in: query
          description: Code of the collection of the specimens (case insensitive)
          schema:
            type: string
          example: CHE
        - name: family
          required: false
          in: query
          description: Family of the specimens (case insensitive)
          schema:
            type: string
          example: Blenniidae
        - name: genus
          required: false
          in: query
          description: Genus of the specimens (case insensitive)
          schema:
            type: string
          example: Salaria
        - name: specific_epithet
          required: false
          in: query
          description: Specific epithet of the specimens (case insensitive)
          schema:
            type: string
          example: basilisca
        - name: start_date
          required: false
          in: query
          description: Minimum date of collect (ISO 8601 YYYY-MM-DD)
          schema:
            type: string
          example: 2023-12-18
        - name: end_date
          required: false
          in: query
          description: Maximum date of collect (ISO 8601 YYYY-MM-DD)
          schema:
            type: string
        - name: collector
          required: false
          in: query
          description: Person who collected the specimen
          schema:
            type: string
          example: 2023-12-18
        - name: continent
          required: false
          in: query
          description: Query specimen where continent field contains this value (case insensitive)
          schema:
            type: array
            items:
              type: string
        - name: country
          required: false
          in: query
          description: Query specimen where country field contains this value (case insensitive)
          schema:
            type: string
        - name: nominative_collection
          required: false
          in: query
          description: Query specimen where nominative collection field contains this value (case insensitive)
          schema:
            type: string
        - name: import_id
          required: false
          in: query
          description: Query specimen for a specific import
          schema:
            type: string
            format: uuid
        - name: storage_name
          required: false
          in: query
          description: Query specimen where storage name field contains this value (case insensitive)
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecimenIntegrationPageResponse'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
    patch:
      operationId: updateMultipleSpecimen
      tags:
        - Specimen
      summary: update multiple specimen
      parameters:
        - name: id
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
              format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecimenIntegrationMergeRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uuid
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
    post:
      operationId: addSpecimen
      tags:
        - Specimen integration
        - Specimen determination
      summary: Specimen integration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecimenIntegrationRequest'
      responses:
        '201':
          description: Successful operation
          headers:
            specimenId:
              description: Specimen Identifier created
              required: true
              schema:
                type: string
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/specimens/nominative-collection/autocomplete:
    get:
      operationId: getNominativeCollections
      tags:
        - Specimen
      summary: Retrieves specimen nominative collections
      parameters:
        - name: q
          in: query
          description: Prefix of the nominative collection
          required: true
          schema:
            type: string
        - name: size
          in: query
          description: Maximum number of nominative collections to return
          required: false
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: A list of nominative collections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoCompleteResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /v1/specimens/storage-name/autocomplete:
    get:
      operationId: getStorageNames
      tags:
        - Specimen
      summary: Retrieves specimen storage names
      parameters:
        - name: q
          in: query
          description: Prefix of the storage name
          required: true
          schema:
            type: string
        - name: size
          in: query
          description: Maximum number of items to return
          required: false
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: A list of string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoCompleteResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
...
components:
  securitySchemes:
    collection-oidc: # <--- Arbitrary name for the security scheme. Used to refer to it from elsewhere.
      type: openIdConnect
      openIdConnectUrl: http://sample.com/realms/recolnat-oauth2/.well-known/openid-configuration
  schemas:
...
    SpecimenIntegrationRequest:
      description: Specimen integration data request
      type: object
      nullable: false
      properties:
        institutionId:
          type: string
          format: uuid
          example: 22b05ae9-25b6-4880-b4ca-52b91bf6bdf6
        collectionId:
          type: string
          format: uuid
          example: 22b05ae9-25b6-4880-b4ca-52b91bf6bdf6
        catalogNumber:
          type: string
        recordNumber:
          type: string
        basisOfRecord:
          type: string
        preparations:
          type: string
        preparationDetail:
          type: string
        individualCount:
          type: string
        approximateIndividualCount:
          type: boolean
          nullable: false
        sex:
          type: string
        lifeStage:
          type: string
        occurrenceRemarks:
          type: string
        legalStatus:
          type: string
        donor:
          type: string
        state:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        nominativeCollection:
          type: string
        identifications:
          type: array
          items:
            $ref: '#/components/schemas/Identification'
        geologicalContext:
          $ref: '#/components/schemas/GeologicalContext'
        collectionEvent:
          $ref: '#/components/schemas/CollectionEvent'
        literatures:
          type: array
          items:
            $ref: '#/components/schemas/Literature'
        medias:
          type: array
          items:
            $ref: '#/components/schemas/Media'
        other:
          $ref: '#/components/schemas/Other'
        management:
          $ref: '#/components/schemas/Management'
    SpecimenIntegrationMergeRequest:
      description: Specimen integration data request
      type: object
      nullable: false
      properties:
        catalogNumber:
          type: string
        recordNumber:
          type: string
        basisOfRecord:
          type: string
        preparations:
          type: string
        individualCount:
          type: string
        sex:
          type: string
        lifeStage:
          type: string
        occurrenceRemarks:
          type: string
        legalStatus:
          type: string
        donor:
          type: string
        geologicalContext:
          $ref: '#/components/schemas/GeologicalContext'
        collectionEvent:
          $ref: '#/components/schemas/CollectionEvent'
...
    SpecimensListResponse:
      description: Specimen integration data response
      type: object
      nullable: false
      properties:
        id:
          type: string
          nullable: false
          example: f4b71ce1-9f8a-4dca-af89-4b61d6ec6884
        collectionId:
          type: string
          nullable: false
          example: Mammals
        scientificName:
          type: string
        scientificNameAuthorship:
          type: string
        collectionName:
          type: string
        catalogNumber:
          type: string
        state:
          type: string
        modifiedAt:
          type: string
          format: date-time
    SpecimenIntegrationPageResponse:
      description: specimen pagination response
      type: object
      properties:
        totalPages:
          type: string
          format: integer
        numberOfElements:
          type: string
          format: integer
        specimenListResponse:
          type: array
          items:
            $ref: '#/components/schemas/SpecimensListResponse'
...
    OperationType:
      type: string
      enum:
        - VALID
        - REVIEW
        - DRAFT
        -
...
    Identification:
      type: object
      description: Identification description
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        errorMessage:
          type: string
          description: Message
        verbatimIdentification:
          type: string
          description: Verbatim Identification
        identificationVerificationStatus:
          type: boolean
          description: Identification Verification Status
        identifiedByID:
          type: string
          description: Persona who identify
        dateIdentified:
          type: string
          description: date sous forme de localtime ou de periode
        currentDetermination:
          type: boolean
          description: Current Determination
        typeStatus:
          type: string
          description: typeStatus
        identificationRemarks:
          type: string
          description: Identification Remarks
        taxon:
          type: array
          items:
            $ref: "#/components/schemas/Taxon"
    GeologicalContext:
      description: GeologicalContext data
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        verbatimEpoch:
          type: string
        ageAbsolute:
          type: string
        range:
          type: string
        earliestAgeOrLowestStage:
          type: string
        latestAgeOrHighestStage:
          type: string
        earliestEpochOrLowestSeries:
          type: string
        latestEpochOrHighestSeries:
          type: string
        earliestPeriodOrLowestSystem:
          type: string
        latestPeriodOrHighestSystem:
          type: string
        earliestEraOrLowestErathem:
          type: string
        latestEraOrHighestErathem:
          type: string
        earliestEonOrLowestEonothem:
          type: string
        latestEonOrHighestEonothem:
          type: string
        lowestBiostratigraphicZone:
          type: string
        highestBiostratigraphicZone:
          type: string
        group:
          type: string
        formation:
          type: string
        member:
          type: string
        bed:
          type: string
        otherLithostratigraphicTerms:
          type: string
    Location:
      type: object
      properties:
        municipality:
          type: string
        stateProvince:
          type: string
        county:
          type: string
        country:
          type: string
        continent:
          type: string
        countryCode:
          type: string
        waterBody:
          type: string
        islandGroup:
          type: string
        island:
          type: string
        locality:
          type: string
        locationRemarks:
          type: string
    Literature:
      description: publication, bibliographic reference, global unique identifier, URI
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        identifier:
          type: string
        authors:
          type: string
        date:
          type: string
        title:
          type: string
        review:
          type: string
        volume:
          type: string
        number:
          type: string
        pages:
          type: string
        publisher:
          type: string
        publicationPlace:
          type: string
        editors:
          type: string
        bookTitle:
          type: string
        pageNumber:
          type: string
        citation:
          type: string
        language:
          type: string
        keywords:
          type: string
        description:
          type: string
        url:
          type: string
        remarks:
          type: string
    Media:
      description: images attached
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        creator:
          type: string
        contributor:
          type: string
        license:
          type: string
        source:
          type: string
        description:
          type: string
        fileName:
          type: string
          nullable: false
        mediaUrl:
          type: string
          description: media url, file URI without server name
          nullable: false
        isCover:
          type: boolean
          description: if this media is a cover or not
    Other:
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        linkGerBank:
          type: string
        linkBold:
          type: string
        linkOther:
          type: string
        financialAid:
          type: string
        computerizationProgram:
          type: string
    Management:
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        storageName:
          type: string
    Taxon:
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        scientificName:
          type: string
          example: Pica pica
          nullable: false
        scientificNameAuthorship:
          type: string
        kingdom:
          type: string
          example: Animalia
        phylum:
          type: string
          example: chordata
        taxonClass:
          type: string
        taxonOrder:
          type: string
        subOrder:
          type: string
        family:
          type: string
        subFamily:
          type: string
        genus:
          type: string
          example: Pica
        subGenus:
          type: string
        specificEpithet:
          type: string
        infraspecificEpithet:
          type: string
        vernacularName:
          type: string
          description: vernacular Name
          example: le poisson rouge
        taxonRemarks:
          type: string
          description: taxon Remarks
          example: this name is a misspelling in common use
        referentialTaxonId:
          type: string
        referentialName:
          type: string
          description: referentialName
        referentialVersion:
          type: string
          description: referentialVersion
        levelType:
          type: string
          enum:
            - MASTER
            - SECONDARY
          description: level Type in hierarchy
    CollectionEvent:
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        noCollectionInformation:
          type: boolean
        eventDate:
          type: string
        interpretedDate:
          type: boolean
        recordedBy:
          type: string
        fieldNumber:
          type: string
        fieldNotes:
          type: string
        eventRemarks:
          type: string
        verbatimLocality:
          type: string
        sensitiveLocation:
          type: boolean
        minimumElevationInMeters:
          type: number
          format: double
        maximumElevationInMeters:
          type: number
          format: double
        interpretedAltitude:
          type: boolean
        minimumDepthInMeters:
          type: number
          format: double
        maximumDepthInMeters:
          type: number
          format: double
        interpretedDepth:
          type: boolean
        decimalLatitude:
          type: number
          format: double
        decimalLongitude:
          type: number
          format: double
        geodeticDatum:
          type: string
        georeferenceSources:
          type: string
        habitat:
          type: string
        location:
          $ref: '#/components/schemas/Location'
...
    AutoCompleteResponse:
      type: array
      items:
        type: string
    BadRequest:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
          example:
            code: 400
            message: "Invalid query parameter"
    NotFound:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
          example:
            code: 404
            message: "Resource not found"
...
    UnexpectedError:
      $ref: '#/components/schemas/Error'
    Forbidden:
      $ref: '#/components/schemas/Error'
    Unauthorized:
      $ref: '#/components/schemas/Error'
    Conflict:
      $ref: '#/components/schemas/Error'
    Error:
      type: object
      description: A common error response
      properties:
        errorCode:
          type: integer
          description: Error Code
        errorMessage:
          type: string
          description: Message
        errorLevel:
          type: string
          description: Level of error
          default: info
          enum:
            - info
            - warning
            - error
            - critical
        errorType:
          type: string
          description: Type of error
          enum:
            - functional
            - technical
        documentationUrl:
          type: string
          description: Url of the documenation
        tips:
          type: string
          description: Tips/help for adressing the error
...